requirements.txt
praw
python-dotenv
langchain_chroma
langchain
langchain_core
langchain_community
langchain_text_splitters
langchain_huggingface
pathlib